refactor: standardize file naming to index.tsx for simpler imports
