+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import { RouterProvider } from "react-router-dom";
-import { router } from "./router.tsx";
+import { router } from "@/router";
 import { QueryClient, QueryClientProvider } from "react-query";
 
 const client = new QueryClient();
